+++ adapters/react-native/ios/LevixelModule.swift
@@ -16,6 +16,9 @@ public final class LevixelModule: Module {
             Prop("galleryId") { (view: LevixelView, galleryId: String?) in
                 view.galleryId = galleryId ?? ""
             }
+            Prop("sourceCornerRadius") { (view: LevixelView, sourceCornerRadius: Double?) in
+                view.sourceCornerRadius = CGFloat(sourceCornerRadius ?? 0)
+            }
             Prop("theme") { (view: LevixelView, theme: LevixelTheme?) in
                 view.theme = theme ?? .dark
             }